# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: audit — API fiscal (ESAA-Flow)
  version: 1.0.0
  description: |
    Contrato mínimo para destravar o frontend (Lovable). Toda escrita passa pelo
    orquestrador single-writer: o cliente envia uma *intenção*; a API valida
    (7 camadas), grava o evento e re-projeta. Leituras vêm da projeção.

    Convenções:
    - Identificadores em inglês; mensagens de erro em PT-BR.
    - `cnpj` sempre com 14 dígitos, sem máscara.
    - `period` no formato `YYYY-MM` (competência).
    - Toda resposta de escrita devolve o `event_seq` gravado e o `projection_hash`.
servers:
  - url: http://localhost:3000/v1
    description: desenvolvimento
  - url: https://api.SEU-DOMINIO.com.br/v1
    description: produção
security:
  - bearerAuth: []
...
paths:
...
  /clients/{cnpj}/sped:
    parameters:
      - $ref: "#/components/parameters/Cnpj"
    post:
      tags: [dossier]
      operationId: importSped
      summary: Importa a EFD-Contribuições do cliente
      description: >
        Síncrono, e `201` em vez de `202` com job: o arquivo é o que o cliente já
        gera hoje, e é a fonte de verdade do saldo credor de PIS/Cofins.


        Lê os registros de que o dossiê precisa — `0000`, `C100`, `C170`, `M100`,
        `M500`, `1100` e `1500` — e conta os demais sem interpretá-los. Um parser
        que anunciasse ler o layout inteiro e ignorasse metade em silêncio seria
        pior do que um que diz o que lê.


        **Versão de layout desconhecida é recusada**, não lida no palpite: entre
        versões os campos mudam de posição, e adivinhar trocaria base por valor.


        O CNPJ do arquivo é conferido contra o do escopo antes de qualquer
        gravação. Divergência devolve `tenant_violation`: importar a EFD de um
        CNPJ na conta de outro produziria um dossiê que acusa o cliente errado.


        Uma escrituração por competência: a retificadora substitui a original,
        como na própria EFD. Manter as duas daria dois saldos credores para o
        mesmo mês e nenhum defensável.
      requestBody:
        required: true
        content:
          text/plain:
            schema: { type: string }
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Escrituração importada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SpedImportOutcome" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Rejected" }
...
  /clients/{cnpj}/efd-icms-ipi:
    parameters:
      - $ref: "#/components/parameters/Cnpj"
    post:
      tags: [reconciliation]
      operationId: importEfdIcmsIpi
      summary: Importa a EFD ICMS/IPI do cliente
      description: >
        Rota própria, e não a `/sped` da EFD-Contribuições. São duas escriturações
        com posições de campo próprias, e um endpoint que adivinhasse qual é pelo
        conteúdo erraria em silêncio no caso que importa: arquivo do tipo errado
        lido com as posições do outro devolve base no lugar de valor.


        Onde os dois leiautes divergem é no `0000` — `DT_INI` no campo 4 contra 6,
        `CNPJ` no 7 contra 9. `C100` e `C170` são idênticos nos dois, o que torna
        a confusão silenciosa e a recusa pelo cabeçalho obrigatória.


        Lê `0000`, `C100`, `C170`, `C190`, `E110` e `E520`, e conta os demais sem
        interpretá-los. **Só o leiaute 020** é aceito, que é o obrigatório desde
        01/01/2026 e o documentado pelo Guia Prático 3.2.2, contra o qual as
        posições foram conferidas uma a uma.


        Convive com a EFD-Contribuições da mesma competência: são escriturações
        diferentes, e uma não substitui a outra. Dentro de cada uma, a
        retificadora substitui a original.
      requestBody:
        required: true
        content:
          text/plain:
            schema: { type: string }
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Escrituração importada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EfdIcmsImportOutcome" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Rejected" }
  /clients/{cnpj}/icms-ipi-reconciliation/{period}:
    parameters:
      - $ref: "#/components/parameters/Cnpj"
      - $ref: "#/components/parameters/PeriodPath"
    get:
      tags: [reconciliation]
      operationId: getIcmsIpiReconciliation
      summary: Confere a EFD ICMS/IPI contra as regras do próprio guia
      description: |
        O arquivo declara a apuração no `E110` e no `E520` e declara os documentos
        que a sustentam nos blocos C. Nada garante que os dois concordem: o PVA
        aceita arquivo cuja apuração não fecha com os documentos, e quem recebe a
        conta é o contribuinte, meses depois.

        Refaz a aritmética que o guia manda o arquivo obedecer e soma os
        documentos contra o declarado. Não é opinião nossa sobre quanto deveria
        ser: é a escrituração conferida contra as regras dela.

        **Conferência que não pôde rodar vem `not_verified` com motivo, nunca
        `passed`.** Sem `E110` não há apuração declarada para conferir, o que não
        é o mesmo que apuração correta; documento sem `C170` não é documento
        errado, porque a NF-e de emissão própria costuma vir só com `C100` e
        `C190`; e quando o arquivo traz `C500`, `D100` e afins, que também lançam
        ICMS e este leitor ainda não soma, comparar só os `C190` acusaria uma
        diferença que é limitação nossa.

        A conciliação é **derivada na leitura**, e não um resultado guardado:
        congelada, uma conferência nova valeria só para arquivo importado depois
        dela, e o cliente continuaria vendo o veredito velho sobre o mesmo
        arquivo.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IcmsIpiReconciliation" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /clients/{cnpj}/fisco-assessments/{period}:
    parameters:
      - $ref: "#/components/parameters/Cnpj"
      - $ref: "#/components/parameters/PeriodPath"
    get:
      tags: [reconciliation]
      operationId: getFiscoComparison
      summary: Contra-apuração — nossa apuração × proposta do Fisco, nota a nota
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FiscoComparison" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [reconciliation]
      operationId: uploadFiscoAssessment
      summary: Envia a proposta do Fisco e compara (emite `assessment.compared`)
      description: >
        O layout do CSV é **nosso**, não da RFB: o formato oficial de exposição da
        apuração assistida ainda está em piloto. Duas formas de envio, o corpo cru
        como `text/csv` e o `multipart/form-data` com um arquivo.


        Duas colunas decidem o modo. Com `chave_acesso`, a proposta é nota a nota
        e a comparação acontece; sem ela, é só totais por tributo, e
        `line_level: false` diz que a comparação nota a nota **não foi feita** —
        zero divergência de item ali não significa que as notas conferem.


        Cabeçalho nota a nota:
        `chave_acesso;item;tributo;sentido;base;aliquota;valor`.
        Cabeçalho só de totais: `tributo;valor`.


        Exige a competência apurada: sem a nossa apuração não há o que comparar,
        e registrar a proposta sozinha daria a impressão de conferência que não
        aconteceu (`422`, camada 4).


        Uma proposta nova substitui a anterior por inteiro, inclusive as
        divergências — manter as duas daria dois números para a mesma competência
        e nenhum deles defensável.
      requestBody:
        required: true
        content:
          text/csv:
            schema: { type: string }
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Proposta registrada e comparada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FiscoComparisonOutcome" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Rejected" }
...
  /deadlines:
    get:
      tags: [reconciliation]
      operationId: getPortfolioCalendar
      summary: Calendário da carteira — prazos datados e pendências, em listas separadas
      description: >
        As duas listas são separadas de propósito. Um **prazo** tem data e, quando
        `nature: normativo`, base legal: perdê-lo tem consequência jurídica. Uma
        **pendência** sai do estado do sistema (competência encerrada e não
        apurada, proposta do Fisco sem resposta) e não tem data de lei nenhuma.
        Dar a mesma cara às duas faria o escritório tratar um alvo interno como
        prazo de norma.


        `normative_rules_loaded: false` significa que nenhum prazo normativo está
        carregado — e portanto que a lista vazia de prazos não é "nada a vencer".
        A tabela nasce vazia porque as datas citadas por terceiros no briefing não
        foram conferidas em texto oficial.
      parameters:
        - in: query
          name: horizon_days
          description: Janela em dias. Prazo já vencido aparece com `days_left` negativo.
          schema: { type: integer, minimum: 1, maximum: 365, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortfolioCalendar" }

  # ---------------------------------------------------------- simulation
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cnpj:
      in: path
      name: cnpj
      required: true
      # Doze posições alfanuméricas e dois dígitos verificadores numéricos: a
      # Receita emite CNPJ alfanumérico desde 31/07/2026 (IN RFB 2.229/2024).
      # O dígito verificador não é conferido pelo schema — JSON Schema não faz
      # módulo 11 —, e sim na fronteira de entrada da API.
      schema: { type: string, pattern: "^[0-9A-Za-z]{12}[0-9]{2}$" }
...
    PeriodPath:
      in: path
      name: period
      required: true
      schema: { $ref: "#/components/schemas/PeriodId" }
...
  responses:
...
    Forbidden:
      description: Papel sem permissão, ou recurso não incluído no plano
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Rejected:
      description: Intenção rejeitada pelo pipeline (evento `output.rejected` gravado)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Rejection" }
  schemas:
    PeriodId:
      type: string
      pattern: "^[0-9]{4}-(0[1-9]|1[0-2])$"
      example: "2027-01"
    Regime:
      type: string
      enum: [mei, simples_integrado, simples_hibrido, lucro_presumido, lucro_real]
...
    RejectionReason:
      type: string
      enum:
        - schema_violation
        - unknown_code
        - code_incompatible
        - regime_violation
        - closed_period_violation
        - fisco_mismatch
        - duplicate_document
        - sequence_gap
        - verification_mismatch
        - boundary_violation
...
    Deadline:
      type: object
      required: [cnpj, kind, name, due_date, days_left, severity, nature, state]
      properties:
        cnpj: { type: string }
        legal_name: { type: string, nullable: true }
        regime: { $ref: "#/components/schemas/Regime" }
        period: { type: string, nullable: true }
        kind: { type: string, example: certificado_a1_vencendo }
        name: { type: string }
        due_date: { type: string, format: date }
        days_left: { type: integer, description: "Negativo quando o prazo já venceu." }
        severity: { type: string, enum: [low, medium, high, critical] }
        nature:
          type: string
          enum: [normativo, fato]
          description: >
            `normativo` é prazo de norma e sempre traz `legal_basis`. `fato` é data
            que o sistema conhece — a validade de um certificado A1, por exemplo —
            e não tem base legal porque não é prazo de lei.
        legal_basis: { type: string, nullable: true }
        state: { type: string, enum: [pending, met, missed, dismissed] }
    Pendency:
      type: object
      required: [kind, cnpj, period, name, message, severity, openSince, daysOpen]
      properties:
        kind:
          type: string
          enum:
            - competencia_nao_apurada
            - apuracao_nao_confirmada
            - proposta_do_fisco_sem_resposta
            - proposta_do_fisco_sem_detalhe
        cnpj: { type: string }
        period: { $ref: "#/components/schemas/PeriodId" }
        name: { type: string }
        message: { type: string }
        severity: { type: string, enum: [low, medium, high, critical] }
        openSince:
          type: string
          format: date-time
          description: Substitui a data de vencimento, que para uma pendência não existe.
        daysOpen: { type: integer }
    PortfolioCalendar:
      type: object
      required: [horizon_days, normative_rules_loaded, deadlines, pendencies]
      properties:
        horizon_days: { type: integer }
        normative_rules_loaded: { type: boolean }
        deadlines:
          type: array
          items: { $ref: "#/components/schemas/Deadline" }
        pendencies:
          type: array
          items: { $ref: "#/components/schemas/Pendency" }
    Divergence:
      type: object
      required:
        [scope, subject, tax, ourCents, fiscoCents, differenceCents, probableCause, severity]
      properties:
        scope: { type: string, enum: [tributo, documento, item] }
        subject:
          type: string
          description: Chave de acesso, `chave#item` ou o nome do tributo.
        tax: { type: string }
        direction: { type: string, enum: [inbound, outbound], nullable: true }
        accessKey: { type: string, nullable: true }
        line: { type: integer, nullable: true }
        ourCents: { type: integer }
        fiscoCents: { type: integer }
        differenceCents:
          type: integer
          description: Fisco menos nosso. Positivo = o Fisco aponta mais do que escrituramos.
        probableCause:
          type: string
          description: >
            Hipótese, não diagnóstico: o sistema vê duas listas de números, e a
            razão real pode ser erro nosso, erro do Fisco, documento cancelado ou
            nota ainda não processada por um dos lados.
          enum:
            - aliquota_divergente
            - base_divergente
            - base_e_aliquota_divergentes
            - valor_incoerente_com_base_e_aliquota
            - debito_nao_escriturado
            - credito_nao_aproveitado
            - debito_nao_reconhecido_pelo_fisco
            - credito_glosado
            - arredondamento
            - total_nao_explicado_pelas_linhas
            - causa_nao_determinada
        severity: { type: string, enum: [low, medium, high, critical] }
    ComparisonSummary:
      type: object
      required:
        [divergencesCount, exposureCents, creditLossCents, creditAtRiskCents,
         bySeverity, linesCompared, lineLevel]
      properties:
        divergencesCount:
          type: integer
          description: Total exato, mesmo quando a lista de itens vem cortada em 500.
        exposureCents:
          type: integer
          description: Débito que o Fisco aponta e não escrituramos. Será cobrado.
        creditLossCents:
          type: integer
          description: Crédito que o Fisco reconhece e não aproveitamos. Dinheiro na mesa.
        creditAtRiskCents:
          type: integer
          description: Crédito nosso que o Fisco não reconhece. Tende a ser glosado.
        bySeverity:
          type: object
          properties:
            critical: { type: integer }
            high: { type: integer }
            medium: { type: integer }
            low: { type: integer }
        linesCompared: { type: integer, description: "Zero quando a proposta veio só com totais." }
        lineLevel: { type: boolean }
    FiscoComparison:
      type: object
      required:
        [period, source, reference, line_level, lines_count, totals,
         uploaded_at, event_seq, divergences, summary]
      properties:
        period: { $ref: "#/components/schemas/PeriodId" }
        source: { type: string, enum: [manual_upload, official_api] }
        reference:
          type: string
          description: >
            Nome do arquivo ou protocolo. É o que permite ao contador dizer depois
            qual proposta recebeu e quando.
        line_level: { type: boolean }
        lines_count: { type: integer }
        totals:
          type: object
          additionalProperties: { type: integer }
        uploaded_at: { type: string, format: date-time }
        event_seq: { type: integer }
        divergences:
          type: array
          maxItems: 500
          items: { $ref: "#/components/schemas/Divergence" }
        summary: { $ref: "#/components/schemas/ComparisonSummary" }
    FiscoComparisonOutcome:
      allOf:
        - $ref: "#/components/schemas/FiscoComparison"
        - type: object
          properties:
            projection_hash: { type: string }
            rejected:
              type: array
              description: Linhas do CSV recusadas. Uma linha ruim não derruba as boas.
              items:
                type: object
                properties:
                  row: { type: integer, description: "Número da linha, contando o cabeçalho como 1." }
                  reason: { type: string }
                  content: { type: string }
...
    SpedImportOutcome:
      type: object
      required:
        [sped_file_id, period, kind, layout_version, documents_count,
         carried_credits_count, rejected, counts, event_seq, projection_hash]
      properties:
        sped_file_id: { type: string, format: uuid }
        period: { $ref: "#/components/schemas/PeriodId" }
        kind: { type: string, enum: [original, retificadora] }
        layout_version: { type: string }
        company_name: { type: string }
        documents_count: { type: integer }
        carried_credits_count: { type: integer }
        rejected:
          type: array
          description: Registros recusados, com a linha e o motivo. Um ruim não derruba os bons.
          items:
            type: object
            properties:
              line: { type: integer }
              record: { type: string }
              reason: { type: string }
        counts:
          type: object
          description: Registros lidos por tipo, para conferir o que entrou.
          additionalProperties: { type: integer }
        event_seq: { type: integer }
        projection_hash: { type: string }
    EfdIcmsImportOutcome:
      type: object
      required:
        [sped_file_id, period, kind, layout_version, uf, documents_count,
         has_icms_assessment, has_ipi_assessment, rejected, counts, event_seq,
         projection_hash]
      properties:
        sped_file_id: { type: string, format: uuid }
        period: { $ref: "#/components/schemas/PeriodId" }
        kind: { type: string, enum: [original, retificadora] }
        layout_version: { type: string }
        company_name: { type: string }
        uf:
          type: string
          description: A EFD ICMS/IPI é estadual; sem a UF não se sabe qual legislação aplicar.
        documents_count: { type: integer }
        has_icms_assessment:
          type: boolean
          description: >
            Se o arquivo trouxe `E110`. Distingue ausência de apuração de apuração
            zerada, que não são a mesma coisa.
        has_ipi_assessment: { type: boolean }
        rejected:
          type: array
          description: Registros recusados, com a linha e o motivo. Um ruim não derruba os bons.
          items:
            type: object
            properties:
              line: { type: integer }
              record: { type: string }
              reason: { type: string }
        counts:
          type: object
          description: Registros lidos por tipo, para conferir o que entrou.
          additionalProperties: { type: integer }
        event_seq: { type: integer }
        projection_hash: { type: string }
    IcmsIpiCheck:
      type: object
      required: [checkId, name, rule, status, severity, issues]
      properties:
        checkId: { type: string }
        name: { type: string }
        rule:
          type: string
          description: A regra conferida, com a origem no guia. Vai para a tela e para o Book.
        status: { type: string, enum: [passed, failed, not_verified] }
        severity: { type: string, enum: [low, medium, high, critical] }
        notVerifiedReason:
          type: string
          nullable: true
          description: >
            Por que não deu para conferir; `null` quando conferiu. `not_verified`
            sem motivo é indistinguível de desleixo — quem lê precisa saber se
            falta arquivo, falta registro ou falta suporte nosso.
        declaredCents: { type: integer, nullable: true }
        expectedCents: { type: integer, nullable: true }
        differenceCents: { type: integer, nullable: true }
        issues:
          type: array
          items:
            type: object
            properties:
              subject: { type: string }
              message: { type: string }
              differenceCents: { type: integer }
    IcmsIpiReconciliation:
      type: object
      required:
        [period, kind, layout_version, imported_at, checks, totalDifferenceCents,
         failedCount, notVerifiedCount]
      properties:
        period: { $ref: "#/components/schemas/PeriodId" }
        kind: { type: string, enum: [original, retificadora] }
        layout_version: { type: string }
        imported_at: { type: string, format: date-time }
        checks:
          type: array
          items: { $ref: "#/components/schemas/IcmsIpiCheck" }
        totalDifferenceCents:
          type: integer
          description: Soma das diferenças absolutas só das conferências que falharam.
        failedCount: { type: integer }
        notVerifiedCount:
          type: integer
          description: >
            Quantas não puderam ser conferidas. Aparece ao lado das que falharam
            para que "nenhuma falha" nunca seja lido como "tudo conferido".
...
    Rejection:
      type: object
      properties:
        rejected: { type: boolean, enum: [true] }
        filename:
          type: string
          description: |
            Arquivo que foi recusado. Presente nas rejeições do `207` de
            ingestão — sem ele, num lote de 200 XMLs a tela sabe que houve
            recusa e não sabe de qual arquivo.
        layer: { type: integer, minimum: 1, maximum: 7 }
        reason: { $ref: "#/components/schemas/RejectionReason" }
        message: { type: string }
        details: { type: object }
        rejection_event_seq: { type: integer, description: Seq do `output.rejected` gravado }
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }